requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==3.2.8
django-braces==1.14.0
django-crispy-forms==1.13.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
Markdown==3.3.4
psycopg2-binary==2.9.1
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
twilio==7.1.0
urllib3==1.26.7
